renamed files within junoshell
